Phase C-1 cascade: move ResumeValueSource + ResumeValueLayoutSummary chain to majit-backend / majit-ir
ResumeValueSource (resume.py:96-188 tagged value source) and
ResumeValueLayoutSummary (the serialization-friendly summary) and
their conversion impls (ResumeValueLayoutSummary <-> ResumeValueSource
<-> ExitValueSourceLayout) form a tight cycle that must move
together; partial-hop migration fails the cross-crate orphan rule.

- ResumeValueLayoutSummary moves to majit-ir::resumedata (depends
  only on Const + Type + ResumeValueKind, all in majit-ir).
- ResumeValueSource moves to majit-backend::resume_value (depends on
  Const and ExitValueSourceLayout, both visible from majit-backend).
- ResumeValueLayoutSummary conversion methods (raw_fail_arg_position,
  to_resume_source, to_exit_source) move to majit-backend as the
  ResumeValueLayoutSummaryExt trait — Rust's orphan rule forbids
  inherent impls on a foreign type, so the impls become a trait
  + impl that callers `use majit_backend::ResumeValueLayoutSummaryExt`
  to bring into scope.
- ResumeValueLayoutSummary::from_exit_value_source becomes a free
  function resume_value_layout_summary_from_exit_value_source.

majit-metainterp re-exports ResumeValueSource and
ResumeValueLayoutSummaryExt for caller compatibility.  Preparatory
step toward backend struct deletion; ResumeData migration continues
this chain.
